Win_API 제거

+++ MinhoEngine_SOURCE/Mininput.cpp
@@ -95,13 +95,13 @@ namespace min {
 		mMousePosition.x = -1.0f;
 		mMousePosition.y = -1.0f;
 
-		if (mousePos.x > 0 && mousePos.x < width)
+		if ((UINT)mousePos.x > 0 && (UINT)mousePos.x < width)
 		{
-			mMousePosition.x = mousePos.x;
+			mMousePosition.x = (float)mousePos.x;
 		}
-		if (mousePos.y > 0 && mousePos.y < height)
+		if ((UINT)mousePos.y > 0 && (UINT)mousePos.y < height)
 		{
-			mMousePosition.y = mousePos.y;
+			mMousePosition.y = (float)mousePos.y;
 		}
 	}
 	void input::claerKey()
